pyproject.toml
[project]
name = "boostcamp-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alex Keyes", email = "alex.o.keyes@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]

[project.scripts]
boostcamp-mcp = "boostcamp_mcp.server:mcp.run"
login = "boostcamp_mcp.login:main"

[tool.uv]
dev-dependencies = [
    "mcp[cli]>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
